Added code for google oauth

+++ application.py
@@ -9,8 +9,20 @@ from database_setup import Base, Category, Item, User
 from flask import session as login_session
 import random
 import string
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
 
 app = Flask(__name__)
+
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Item Catalog"
 
 
 # Connect to database
@@ -29,6 +41,171 @@ def showLogin():
                     for x in xrange(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
+
+
+# GOOGLE Connect
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+    # ADD PROVIDER TO LOGIN SESSION
+    login_session['provider'] = 'google'
+
+    # see if user exists, if it doesn't make a new one
+    user_id = getUserID(data["email"])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("You are now logged in as %s" % login_session['username'])
+    print("Done!")
+    return output
+
+
+# GOOGLE Disconnect - Revoke a current user's token and reset their login_session
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    if result['status'] == '200':
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
+# Disconnect based on provider
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['access_token']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('publicCatalog'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('publicCatalog'))
+
+
+# User Helper Functions
+def createUser(login_session):
+    session = connect()
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    session = connect()
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    session = connect()
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
 
 
 # JSON endpoint for catalog categories
@@ -62,13 +239,18 @@ def showCatalog():
     session = connect()
     categories = session.query(Category).all()
     items = session.query(Item).order_by(desc(Item.id)).limit(10).all()
-    return render_template('catalog.html', categories=categories, items=items)
+    if 'username' not in login_session:
+        return render_template('publicCatalog.html', categories=categories, items=items)
+    else:
+        return render_template('catalog.html', categories=categories, items=items)
 
 
 # Add new category
 @app.route('/catalog/new/', methods=['GET', 'POST'])
 def newCategory():
     session = connect()
+    if 'username' not in login_session:
+        return redirect('/login')
     categories = session.query(Category).all()
     if request.method == 'POST':
         category = Category(name = request.form['name'])
@@ -86,6 +268,11 @@ def editCategory(category_id):
     session = connect()
     categories = session.query(Category).all()
     category = session.query(Category).filter_by(id=category_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if category.user_id != login_session['user_id']:
+        return """<script>function myFunction() {alert('You are not authorized to edit this category. 
+            Please create your own category in order to edit.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         if request.form['name']:
             category.name = request.form['name']
@@ -102,6 +289,11 @@ def deleteCategory(category_id):
     session = connect()
     categories = session.query(Category).all()
     category = session.query(Category).filter_by(id=category_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if category.user_id != login_session['user_id']:
+        return """<script>function myFunction() {alert('You are not authorized to delete this category. 
+            Please create your own category in order to delete.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         session.delete(category)
         session.commit()
@@ -118,15 +310,24 @@ def showCategory(category_id):
     session = connect()
     categories = session.query(Category).all()
     category = session.query(Category).filter_by(id=category_id).one()
+    creator = getUserInfo(category.user_id)
     items = session.query(Item).filter_by(category_id=category.id).all()
-    return render_template('showCategory.html', category=category, items=items, categories=categories)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicCategory', category=category, items=items, categories=categories)
+    else:
+        return render_template('showCategory.html', category=category, items=items, categories=categories)
 
 
 # Add a new item
 @app.route('/catalog/<int:category_id>/items/new', methods=['GET', 'POST'])
 def newItem(category_id):
     session = connect()
+    if 'username' not in login_session:
+        return redirect('/login')
     categories = session.query(Category).all()
+    if category.user_id != login_session['user_id']:
+        return """<script>function myFunction() {alert('You are not authorized to create a new item here. 
+            Please create your own category in order to create.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         newItem = Item(name = request.form['name'], description = request.form['description'],
             price = request.form['price'], category_id = category_id)
@@ -142,8 +343,13 @@ def newItem(category_id):
 @app.route('/catalog/<int:category_id>/items/<int:item_id>/edit', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
     session = connect()
+    if 'username' not in login_session:
+        return redirect('/login')
     categories = session.query(Category).all()
     item = session.query(Item).filter_by(category_id = category_id, id = item_id).one()
+    if category.user_id != login_session['user_id']:
+        return """<script>function myFunction() {alert('You are not authorized to edit a new item here. 
+            Please create your own category in order to edit.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         if request.form['name']:
             item.name = request.form['name']
@@ -165,8 +371,13 @@ def editItem(category_id, item_id):
 @app.route('/catalog/<int:category_id>/items/<int:item_id>/delete', methods=['GET', 'POST'])
 def deleteItem(category_id, item_id):
     session = connect()
+    if 'username' not in login_session:
+        return redirect('/login')
     categories = session.query(Category).all()
     item = session.query(Item).filter_by(category_id = category_id, id = item_id).one()
+    if category.user_id != login_session['user_id']:
+        return """<script>function myFunction() {alert('You are not authorized to delete a new item here. 
+            Please create your own category in order to delete.');}</script><body onload='myFunction()'>"""
     if request.method == 'POST':
         session.delete(item)
         session.commit()
